requirements.txt
alembic==1.5.8
asgiref==3.3.4
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Django==3.2
djangorestframework==3.12.4
env==0.1.0
environs==9.3.2
Faker==6.0.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
flask-oidc==1.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
httplib2==0.19.0
idna==2.10
iniconfig==1.1.1
ipdb==0.13.5
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.11.1
mccabe==0.6.1
numpy==1.20.1
oauth2client==4.1.3
okta==0.0.4
packaging==20.8
pandas==1.2.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
playsound==1.2.2
pluggy==0.13.1
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
PTable==0.9.2
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
pyzmq==22.0.2
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.7
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
